+++ CHANGELOG.rst
@@ -7,6 +7,8 @@ Changelog
 * improved RankedPosts class
 * New AccountPosts class for account feed and more
 * Comment class has been cleaned up, use_tags_api has been replaced by api
+* Add check if get_account_votes is still supported (is dropped in HF24)
+* New SupportedByHivemind exception has been added
 
 0.24.11
 -------